requirements.txt
Django==5.0
pillow
pytesseract
djangorestframework
django-crispy-forms
celery
redis
psycopg2-binary
python-dotenv
